moved ActiveModel helper to a separate file
