+++ lib/validates_email_format_of.rb
@@ -137,25 +137,5 @@ module ValidatesEmailFormatOf
   end
 end
 
-if defined?(ActiveModel)
-  class EmailFormatValidator < ActiveModel::EachValidator
-    def validate_each(record, attribute, value)
-      err = ValidatesEmailFormatOf::validate_email_format(value, options)
-      unless err.nil?
-        record.errors[attribute] << err
-        record.errors[attribute].flatten!
-      end
-      #if record.errors.messages[attribute] == [] or record.errors.messages[attribute].nil?
-      #  record.errors.delete(attribute)
-      #end
-    end
-  end
-
-  module ActiveModel::Validations::HelperMethods
-    def validates_email_format_of(*attr_names)
-      validates_with EmailFormatValidator, _merge_attributes(attr_names)
-    end
-  end
-end
-
+require 'validates_email_format_of/active_model' if defined?(::ActiveModel)
 require 'validates_email_format_of/railtie' if defined?(::Rails)
